web/mapplotter/requirements.txt
geopandas
keplergl
palettable
# localtileserver
streamlit
streamlit-folium
streamlit-keplergl
streamlit-bokeh-events
leafmap
django-leaflet

